requirements.txt
discord.py>=2.3.2
wavelink>=3.0.0
python-dotenv
asyncio
